split components.py
